# openapi.generated2.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /api/v1/placeholders/bind:
    post:
      tags:
      - Placeholders
      - Bindings
      summary: Bind placeholder
      description: 'headers_validator: schemas/http_headers.schema.json; Idempotency-Key;
        If-Match; uses schemas/placeholder_probe.schema.json -> schemas/probe_receipt.schema.json;
        returns schemas/bind_result.schema.json'
      operationId: post_placeholders_bind_api_v1_placeholders_bind_post
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - type: object
                additionalProperties: true
              - type: 'null'
              title: Probe
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '428':
          content:
            application/problem+json: {}
          description: Precondition Required
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/placeholders/unbind:
    post:
      tags:
      - Placeholders
      - Bindings
      summary: Unbind placeholder
      description: 'headers_validator: schemas/http_headers.schema.json; Idempotency-Key;
        If-Match; returns schemas/unbind_response.schema.json'
      operationId: post_placeholders_unbind_api_v1_placeholders_unbind_post
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '428':
          content:
            application/problem+json: {}
          description: Precondition Required
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/questions/{id}/placeholders:
    get:
      tags:
      - Placeholders
      - Bindings
      summary: List placeholders by question
      description: returns schemas/list_placeholders_response.schema.json
      operationId: get_question_placeholders_api_v1_questions__id__placeholders_get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      - name: document_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{id}/bindings:purge:
    post:
      tags:
      - Bindings
      - Documents
      summary: Purge bindings for a document
      description: accepts schemas/purge_request.schema.json; returns schemas/purge_response.schema.json
      operationId: post_document_bindings_purge_api_v1_documents__id__bindings_purge_post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            title: Id
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
            title: If-Match
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError